fix: collection paths invalid
