+++ docSite/content/zh-cn/docs/development/upgrading/486.md
@@ -37,3 +37,4 @@ curl --location --request POST 'https://{{host}}/api/admin/initv486' \
 3. 新增 - 网页抓取和数学计算器系统插件
 4. 新增 - 移动文本加工和自定义反馈到基础节点中
 5. 修复 - 工作流中团队插件加载异常
+6. 修复 - 知识库集合目录导航失效
