+++ packages/service/common/mongo/utils.ts
@@ -1,0 +1,4 @@
+export const readFromSecondary = {
+  readPreference: 'secondaryPreferred',
+  readConcern: 'local'
+};
